test: split pmctl tests into smaller tests
Break the pmctl tests up so that if they are slow the whole suite
doesn't timeout and fail.
